flutter_application_1/backend/requirements.txt
flask==3.0.0
stripe==7.8.0
python-dotenv==1.0.0
flask-cors==4.0.0
firebase-admin==6.3.0
mediapipe>=0.10.0
opencv-python
numpy
werkzeug==3.0.3
